+++ ConsoleMarkdownRenderer.Spectre/ObjectRenderers/ConsoleYamlFrontMatterBlockRenderer.cs
@@ -14,28 +14,8 @@ namespace BoxOfYellow.ConsoleMarkdownRenderer.Spectre.ObjectRenderers;
 internal class ConsoleYamlFrontMatterBlockRenderer : ConsoleObjectRendererBase<YamlFrontMatterBlock>
 {
     protected override void Write(ConsoleRenderer renderer, YamlFrontMatterBlock obj)
-    {
-        renderer
+        => renderer
             .NewFrame()
-            .PushStyle(renderer.Options.YamlFrontMatter)
-            .StartInline()
-            .AddInLine("---")
-            .AddInLine(Environment.NewLine);
-
-        for (int i = 0; i < obj.Lines.Lines.Length; i++)
-        {
-            if (!string.IsNullOrEmpty(obj.Lines.Lines[i].Slice.Text))
-            {
-                renderer
-                    .WriteEscape(ref obj.Lines.Lines[i].Slice)
-                    .AddInLine(Environment.NewLine);
-            }
-        }
-
-        renderer
-            .AddInLine("---")
-            .EndInline()
-            .PopStyle()
+            .AddFilledBlock(obj, renderer.Options.YamlFrontMatter, fence: "---")
             .CompleteFrame();
-    }
 }